Http session, UserAuthentication added to all Controlelrs

+++ src/main/java/com/firstyearproject/salontina/Controllers/FrontOfficeControllers/UserAccessController.java
@@ -5,6 +5,7 @@ import com.firstyearproject.salontina.Models.LoginToken;
 import com.firstyearproject.salontina.Models.User;
 import com.firstyearproject.salontina.Services.BookingServiceImpl;
 import com.firstyearproject.salontina.Services.ProductServiceImpl;
+import com.firstyearproject.salontina.Services.UserAuthenticator;
 import com.firstyearproject.salontina.Services.UserServiceImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -43,6 +44,9 @@ public class UserAccessController {
     @Autowired
     BookingServiceImpl bookingService;
 
+    @Autowired
+    UserAuthenticator userAuthenticator;
+
     //Luca
     //Used in Java Methods/mappings
     public void confirmation(String text){
@@ -58,6 +62,7 @@ public class UserAccessController {
         showConfirmation = false;
     }
 
+    //Jonathan & Mike
     private Model userExcists(Model model, HttpSession session) {
         if(session.getAttribute("user") != null) {
             User user = (User)session.getAttribute("user");
@@ -79,7 +84,7 @@ public class UserAccessController {
 
     //Luca
     @PostMapping("/")
-    public String login(Model model, HttpSession session, @ModelAttribute LoginToken loginToken) {
+    public String login(HttpSession session, @ModelAttribute LoginToken loginToken) {
         User user = userService.authenticateUser(loginToken);
 
         if(user != null){
@@ -99,19 +104,28 @@ public class UserAccessController {
     //Jonathan
     @GetMapping("/editUser")
     public String editUser(HttpSession session, Model model) {
+        if(!userAuthenticator.userIsAdmin(session) || !userAuthenticator.userIsUser(session)){
+             return REDIRECT;
+        }
         User user = (User)session.getAttribute("user");
         model.addAttribute("user", user);
         return EDITUSER;
     }
     //Jonathan
     @PostMapping("/editUser")
-    public String editUser(@ModelAttribute User user) {
+    public String editUser(@ModelAttribute User user, HttpSession session) {
+        if(!userAuthenticator.userIsAdmin(session) || !userAuthenticator.userIsUser(session)){
+            return REDIRECT;
+        }
         userService.editUser(user);
         return REDIRECT + USERPROFILE;
     }
 
     @GetMapping("userprofile")
     public String userprofile(Model model, HttpSession session) {
+        if(!userAuthenticator.userIsAdmin(session) || !userAuthenticator.userIsUser(session)){
+            return REDIRECT;
+        }
         User user = (User)session.getAttribute("user");
         model.addAttribute("user", user);
         return USERPROFILE;
@@ -119,7 +133,10 @@ public class UserAccessController {
 
     //Mike
     @GetMapping("/deleteuser/{userid}")
-    public String deleteUser(@PathVariable int userId) {
+    public String deleteUser(@PathVariable int userId, HttpSession session) {
+        if(!userAuthenticator.userIsAdmin(session) || !userAuthenticator.userIsUser(session)){
+            return REDIRECT;
+        }
         userService.deleteUser(userId);
         return REDIRECT + EDITUSER;
     }
@@ -127,6 +144,9 @@ public class UserAccessController {
     //Mike
     @GetMapping("/myprofile")
     public String myprofil(Model model, HttpSession session) {
+        if(!userAuthenticator.userIsAdmin(session) || !userAuthenticator.userIsUser(session)){
+            return REDIRECT;
+        }
         User user = (User)session.getAttribute("user");
         model.addAttribute("user", user);
         return MYPROFILE;
@@ -135,19 +155,16 @@ public class UserAccessController {
     //Mike
     @GetMapping("/contact")
     public String contact(Model model, HttpSession session) {
-        if(session.getAttribute("user") != null) {
-            User user = (User)session.getAttribute("user");
-            model.addAttribute("user", user);
-        } else {
-            model.addAttribute("user", new User());
-            model.addAttribute("loginToken", new LoginToken());
-        }
+        userExcists(model, session);
         return CONTACT;
     }
 
     //Asbjørn
     @PostMapping("subscribeNewsletter")
-    public String subscribeNewsletter (@ModelAttribute User user) {
+    public String subscribeNewsletter (@ModelAttribute User user, HttpSession session) {
+        if(!userAuthenticator.userIsAdmin(session) || !userAuthenticator.userIsUser(session)){
+            return REDIRECT;
+        }
         taskResult = userService.subscribeNewsletter(user.getUserId());
         if (taskResult) {
             confirmation("Du er blevet tilmeldt nyhedsbrevet");
@@ -159,7 +176,10 @@ public class UserAccessController {
 
     //Asbjørn
     @PostMapping("unsubscribeNewsletter")
-    public String unsubscribeNewsletter (@ModelAttribute User user) {
+    public String unsubscribeNewsletter (@ModelAttribute User user, HttpSession session) {
+        if(!userAuthenticator.userIsAdmin(session) || !userAuthenticator.userIsUser(session)){
+            return REDIRECT;
+        }
         taskResult = userService.unsubscribeNewsletter(user.getUserId());
         if (taskResult) {
             confirmation("Du er blevet afmeldt nyhedsbrevet");
@@ -167,11 +187,6 @@ public class UserAccessController {
         }
         confirmation("Vi kunne IKKE afmelde dig nyhedsbrevet. Prøv igen senere");
         return EDITUSER;
-    }
-
-    @GetMapping ("/contact1")
-    public String contact () {
-        return CONTACT;
     }
 
 }
